Move internalServer code into its own source files.
